+++ uefi/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 - Implemented `PartialEq<char>` for `Char8` and `Char16`.
 - Added `CStr16::from_char16_with_nul` and `Char16::from_char16_with_nul_unchecked`.
+- Added terminal GUID constants to `device_path::messaging::Vendor`.
 
 ## Changed
 - `DevicePath::to_string` and `DevicePathNode::to_string` now return